+++ lizard_map/media/lizard_map/lizard_map.js
@@ -204,30 +204,6 @@ function setUpAddWorkspaceItem() {
 
 
 function setUpWorkspaceAcceptable() {
-    // Set up draggability for current and future items.
-    // See http://tinyurl.com/29lg4y3 .
-    $(".workspace-acceptable").live("mouseover", function () {
-        var html;
-        if (!$(this).data("draggable-initialized")) {
-            $(this).data("draggable-initialized", true);
-            $(this).draggable({
-                scroll: 'false',
-                cursor: 'move',
-                helper: 'clone',
-                appendTo: 'body',
-
-                revert: 'true',
-                placeholder: 'ui-sortable-placeholder'
-            });
-        }
-        if (!$(this).data("add-workspace-item-initialized")) {
-            // Add the "add to workspace" button
-            $(this).data("add-workspace-item-initialized", true);
-            html = $(this).html();
-            html = '<span class="ss_sprite ss_add sidebarbox-action-icon add-workspace-item" title="Voeg laag toe aan workspace">&nbsp;</span>' + html;
-            $(this).html(html);
-        }
-    });
     // Do not use mouseout: Mouseout is also activated when moving to
     // the "add-workspace-item" button.
     $(".workspace-acceptable").live("mouseleave", function () {